move server code into pkg
