+++ wailsapp/main.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"github.com/KhoalaS/godel"
+	"github.com/KhoalaS/godel/pkg/server"
 	"github.com/wailsapp/wails/v2"
 	"github.com/wailsapp/wails/v2/pkg/options"
 	"github.com/wailsapp/wails/v2/pkg/options/assetserver"
@@ -10,6 +11,8 @@ import (
 func main() {
 	// Create an instance of the app structure
 	app := NewApp()
+
+	go server.RunServer()
 
 	// Create application with options
 	err := wails.Run(&options.App{
